pyproject.toml
[project]
name = "document-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "inngest>=0.5.18",
    "llama-index-core>=0.14.19",
    "llama-index-readers-file>=0.6.0",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "streamlit>=1.55.0",
    "uvicorn>=0.42.0",
]
